first working implementation of forked streams in TDataFrame2.hpp
TDataFrame simply calls

   while(r.Next())
      for(auto actionPtr : fBookedActions)
         actionPtr->Run(r.GetCurrentEntry());

TDataFrameAction checks the previous filter before executing, and each filter
checks the previous one before applying the filter, in a recursive chain of
calls that ends at TDataFrame.
The trick to not apply the same filter twice is caching the result for
each entry: if a filter is checked a second time for the same entry, it simply
returns the previous result without executing its lambda and without invoking
the previous filters, interrupting the recursive chain of calls early.

test2 shows forked streams usage.