Return smart pointer ActionResultPtr from actions
this pointer is a wrapper around the actual result and lazily triggers
the run of the functional chain upon usage of its value.
All the ActionResultPtr relative to a certain chain are attached to the
original TDataFrame. A pointer is immediately returned when an action
is invoked but no actual calculation takes place.
When TDataFrame::Run is executed the values of these pointers are
filled.
TDataFrame::Run is implicitly called also with the invocation of the
method -> or get.
Therefore, calling -> or get on one pointer causes all of the other
pointers attached to the main chain to be filled with their value.