fix mismatch between booked and returned ActionResultPtr
This commit fixes the issue for which accessing action results multiple
times resulted in spurious triggering of TDataFrame::Run.
TDataFrameInterface was building and booking a TActionResultPtr which
was (necessarily) different from the one returned to the user.
TActionResultPtr::fReady is now encapsulated in a shared pointer so that
objects copy-constructed from the booked TActionResultPtr see the correct
value of fReady.
Type guessing logic has been slightly simplified.
Min, Max and Mean now return a TActionResultPtr<T> (T = type of branch)
instead of always returning a TActionResultPtr<double>. T must fulfill
the requirement of allowing a call to std::numeric_limits<T>.